move ArenaEnemyNPC, update queen slime map area level
