+++ Content/Tiles/BossDomain/UnderworldSpawnerTile.cs
@@ -1,4 +1,4 @@
-﻿using PathOfTerraria.Common.Subworlds.BossDomains.Prehardmode.WoFDomain;
+﻿using PathOfTerraria.Common.Subworlds;
 using PathOfTerraria.Content.NPCs.HellEvent;
 using Terraria.DataStructures;
 using Terraria.Enums;
